+++ docs/changelog.rst
@@ -3,6 +3,13 @@
 What's new
 ##########
 
+v0.1.3 (June 2018)
+--------------------
+
+**Bayesian**
+
+a. T-test now reports the Bayes factor of the alternative hypothesis (BF10)
+
 v0.1.2 (June 2018)
 --------------------